requirements.txt
Scrapy==1.0.4
Pillow==3.1.0
pymongo==2.7.2
Flask==0.10.1
